% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaddata.R
\name{download_data}
\alias{download_data}
\title{Download the Aquamaps database
The location defaults to ~/.config/aquamaps}
\usage{
download_data(force = FALSE)
}
\arguments{
\item{force}{Whether to overwrite an existing database file}
}
\description{
I *think* this can be changed using rappdirs::app_dir("aquamaps")$config()
}
\examples{
download_data()
}
