% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_shapes}
\alias{get_shapes}
\title{Get Aquamaps-based Probability values}
\usage{
get_shapes(
  rds,
  read_cutoff = 10,
  db_file = paste0(rappdirs::app_dir("aquamaps")$config(), "/am.db")
)
}
\arguments{
\item{rds}{A RDS phyloseq object as returned by the ampliseq pipeline}

\item{db_file}{Location of the Aquamaps database. Default is
paste0(rappdirs::app_dir("aquamaps")$config(), '/am.db')}

\item{reads_cutoff}{The number of reads for an ASV to be included in
the analysis. Default is 10.}
}
\description{
Joins the ASV-based sightings data with the Aquamaps
C-squares and returns the Aquamaps-probabilities for each
species.
}
\examples{
shp <- get_shapes("some_examples.rds")
}
