% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_shapes}
\alias{plot_shapes}
\title{Plot Aquamaps-based Probability values}
\usage{
plot_shapes(
  table,
  subset = sample(unique(dplyr::pull(dplyr::filter(table, !is.na(Probability)), ASV)),
    10)
)
}
\arguments{
\item{table}{The large Probabilities table produced by get_shapes()}

\item{subset}{A vector with the names of ASVs to plot. Defaults to 10 at random that have a Probability.}
}
\description{
Takes the shape as produced by get_shapes()
and plots the probabilites for each ASV. A useful helper function.
}
\examples{
plot_shapes(get_shapes("some_examples.rds"))
}
